+++ packages/ui/src/components/Layout.tsx
@@ -2,7 +2,7 @@ import Permission from "@moj-bichard7/common/types/Permission"
 import { useCurrentUser } from "context/CurrentUserContext"
 import { usePathname } from "next/navigation"
 import { useRouter } from "next/router"
-import { LinkButton } from "./Buttons"
+import { LinkButton } from "./Buttons/LinkButton"
 import ConditionalRender from "./ConditionalRender"
 import Header from "./Header"
 import { Banner } from "./Layout.styles"
