+++ packages/ui/src/components/ReallocationNotesForm.tsx
@@ -5,7 +5,7 @@ import { useCsrfToken } from "context/CsrfTokenContext"
 import Link from "next/link"
 import { FormEventHandler, useState } from "react"
 import getForcesForReallocation from "services/getForcesForReallocation"
-import { Button } from "./Buttons"
+import { Button } from "./Buttons/Button"
 import ButtonsGroup from "./ButtonsGroup"
 import Form from "./Form"
 import { NoteTextArea } from "./NoteTextArea"
