+++ packages/ui/src/features/CourtCaseDetails/Tabs/Panels/Notes/AddNoteForm.tsx
@@ -1,4 +1,4 @@
-import { Button } from "components/Buttons"
+import { Button } from "components/Buttons/Button"
 import ConditionalRender from "components/ConditionalRender"
 import { NoteTextArea } from "components/NoteTextArea"
 import { MAX_NOTE_LENGTH } from "config"
