+++ packages/ui/src/features/CourtCaseDetails/Tabs/Panels/Offences/Offence/OffenceNavigation.tsx
@@ -1,4 +1,4 @@
-import { SecondaryButton } from "components/Buttons"
+import { SecondaryButton } from "components/Buttons/SecondaryButton"
 import { BackToAllOffencesLink } from "./BackToAllOffencesLink"
 import { NextButton, PreviousButton } from "./OffenceNavigation.styles"
